Unify the ClientProperties registries
