+++ patches/minecraft/net/minecraft/client/gui/screen/Screen.java.patch
@@ -5,7 +5,7 @@
  
     protected void func_230457_a_(MatrixStack p_230457_1_, ItemStack p_230457_2_, int p_230457_3_, int p_230457_4_) {
 -      this.func_243308_b(p_230457_1_, this.func_231151_a_(p_230457_2_), p_230457_3_, p_230457_4_);
-+      FontRenderer font = net.minecraftforge.client.ItemClientProperties.getFont(p_230457_2_);
++      FontRenderer font = net.minecraftforge.client.ForgeClientProperties.getFont(p_230457_2_);
 +      net.minecraftforge.fml.client.gui.GuiUtils.preItemToolTip(p_230457_2_);
 +      this.renderWrappedToolTip(p_230457_1_, this.func_231151_a_(p_230457_2_), p_230457_3_, p_230457_4_, (font == null ? this.field_230712_o_ : font));
 +      net.minecraftforge.fml.client.gui.GuiUtils.postItemToolTip();
