+++ patches/minecraft/net/minecraft/client/particle/ParticleManager.java.patch
@@ -78,7 +78,7 @@
  
     public void func_180533_a(BlockPos p_180533_1_, BlockState p_180533_2_) {
 -      if (!p_180533_2_.func_196958_f()) {
-+      if (!p_180533_2_.isAir(this.field_78878_a, p_180533_1_) && !net.minecraftforge.client.BlockClientProperties.addDestroyEffects(p_180533_2_, this.field_78878_a, p_180533_1_, this)) {
++      if (!p_180533_2_.isAir(this.field_78878_a, p_180533_1_) && !net.minecraftforge.client.ForgeClientProperties.addDestroyEffects(p_180533_2_, this.field_78878_a, p_180533_1_, this)) {
           VoxelShape voxelshape = p_180533_2_.func_196954_c(this.field_78878_a, p_180533_1_);
           double d0 = 0.25D;
           voxelshape.func_197755_b((p_228348_3_, p_228348_5_, p_228348_7_, p_228348_9_, p_228348_11_, p_228348_13_) -> {
@@ -90,7 +90,7 @@
 +
 +   public void addBlockHitEffects(BlockPos pos, net.minecraft.util.math.BlockRayTraceResult target) {
 +      BlockState state = field_78878_a.func_180495_p(pos);
-+      if (!net.minecraftforge.client.BlockClientProperties.addHitEffects(state, field_78878_a, target, this))
++      if (!net.minecraftforge.client.ForgeClientProperties.addHitEffects(state, field_78878_a, target, this))
 +         func_180532_a(pos, target.func_216354_b());
     }
