+++ patches/minecraft/net/minecraft/client/renderer/ItemRenderer.java.patch
@@ -35,7 +35,7 @@
 +            }
           } else {
 -            ItemStackTileEntityRenderer.field_147719_a.func_239207_a_(p_229111_1_, p_229111_2_, p_229111_4_, p_229111_5_, p_229111_6_, p_229111_7_);
-+            net.minecraftforge.client.ItemClientProperties.getItemStackTileEntityRenderer(p_229111_1_).func_239207_a_(p_229111_1_, p_229111_2_, p_229111_4_, p_229111_5_, p_229111_6_, p_229111_7_);
++            net.minecraftforge.client.ForgeClientProperties.getItemStackTileEntityRenderer(p_229111_1_).func_239207_a_(p_229111_1_, p_229111_2_, p_229111_4_, p_229111_5_, p_229111_6_, p_229111_7_);
           }
  
           p_229111_4_.func_227865_b_();
