+++ src/main/java/net/minecraftforge/client/gui/ForgeIngameGui.java
@@ -51,7 +51,7 @@ import net.minecraft.util.StringUtils;
 import net.minecraft.util.Util;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.GameType;
-import net.minecraftforge.client.ItemClientProperties;
+import net.minecraftforge.client.ForgeClientProperties;
 import net.minecraftforge.client.event.RenderGameOverlayEvent;
 import net.minecraftforge.client.event.RenderGameOverlayEvent.ElementType;
 import net.minecraftforge.common.MinecraftForge;
@@ -293,7 +293,7 @@ public class ForgeIngameGui extends IngameGui
             }
             else
             {
-                ItemClientProperties.renderHelmetOverlay(itemstack, minecraft.player, this.screenWidth, this.screenHeight, partialTicks);
+                ForgeClientProperties.renderHelmetOverlay(itemstack, minecraft.player, this.screenWidth, this.screenHeight, partialTicks);
             }
         }
